requirements.txt
fastapi==0.122.0
uvicorn==0.30.0
python-multipart
asyncpg==0.29.0
sqlalchemy>=2.0.30
pydantic>=2.7.4
pydantic-settings>=2.2.1

PyPDF2>=3.0.1 
python-dotenv>=1.0.1 

python-dotenv
gigachat
pypdf
python-docx
requests
gigachain

